Issue #270: Add control for cpu_model and cpu_feature.
Add configuration parameters for `cpu_model` and `cpu_feature`
which control the setting of the following xml block.

    <cpu mode='custom' match='exact'>
        <model fallback='allow'>core2duo</model>
        <feature policy='require' name='vmx'/>
    </cpu>

This allows a vagrant KVM guest to provide nested virtualization.

NOTE: The `cpu_model` parameter was previously was listed to define
the cpu architecture but it actually was not doing anything. This
parameter has been reused to reflect the preferred guest cpu model.